* Set colorize-fixtures as the folder for extensionHost (debuggin) (#281)
The issue with #174 was the main regex, specially the beginning match, which only matches an expression which ends with a back tick. When you use multiline properties your top line no longer ends the same way (`styled.div<{`). This means we need to set a new beginning match. I tried doing this as a new rule but it just didn't work, instead I split the regex into 2 (non-captured) groups and the second group matches:
```
}>`
```

and sets that as the beginning match, so anything after the above will be treated as SCSS, up until the next backtick

* [before](https://regex101.com/r/yi6jOv/1/)
* [after](https://regex101.com/r/mNfB4E/1/)
* Set colorize-fixtures as the folder for extensionHost (debugging)
* create a multiline fixture from issue #174
* Fixed by creating two groups, one which was the same as before, and one which adds support for }>` syntax as the beginning. Tests have passed